Add mu-plugins loader
The fork is loaded as a mu-plugin, which means that it's hidden from the list of plugins so most users will not even realize the actual plugin is installed.

However, users might want to install the official full-blown version of CoBlocks and that shouldn't cause any errors or issues. The loader implements that by making sure that the mu fork is not loaded if a regular CoBlocks plugin is detected.